# Use Context.context in command execution for funcbench