1. Use context to kill command, it finishes @bwplotka 's TODO
~2. Format the benchmark result, add PR number in the column header, see @geekodour 's comments in #350.~