@geekodour 2020-05-04
I've another format in mind:

Instead of :
| Benchmark | Old ns/op | PR-54 ns/op | Delta |
|-|-|-|-|
BenchmarkBufferedSeriesIterator-8|15.7|15.7|+0.00%

We could also do:
**Old:** `master`/`9a21fdcd1b787b618f1f946e3434391bc5e6271b`
**New:** `PR-3455`
| Benchmark | Old ns/op | New ns/op | Delta |
|-|-|-|-|
BenchmarkBufferedSeriesIterator-8|15.7|15.7|+0.00%

This way we keep the markdown for matter independent of being run on a PR, we can write a wrapper function that adds the **Old** and **New** legends to it.

Would like some feedback on this one.