@nevill 2020-05-04
I like the idea to have a legend.
We can keep current implementation `formatCommentToMD` in `bench.go` for now, though the test being in `main_test.go` is a bit strange.

A better idea is to have a `RenderToMarkDown` in `cmp.go`, like `Render` there. Even implement a `FormatMarkDown`, if we switch to `benchstat`.