@nevill 2020-05-04
Somewhere we mentioned that we need a kind of summary of the current funcbench result, otherwise it's too long to check.
A sample is [here](https://github.com/nevill/prometheus/pull/3#issuecomment-615084625).