@geekodour 2020-05-04
> Somewhere we mentioned that we need a kind of summary of the current funcbench result, otherwise it's too long to check.

@nevill let's keep the changes to the comment to another PR as it seems some more thought will need to be given to the formatting. We could possibly use collapsible sections. (https://github.com/dear-github/dear-github/issues/166#issuecomment-236342209)

> We can keep current implementation `formatCommentToMD` in bench.go for now, though the test being in main_test.go is a bit strange. A better idea is to have a RenderToMarkDown in cmp.go, like Render there.

Agreed with the suggestion for the file change for the test and function `formatCommentToMD` and yes, `RenderToMarkDown` does seem like a better name.