@nevill 2020-05-04
@geekodour I removed the changes to `formatCommentToMD`, it's ready to merge.